add configuration for Meal

+++ src/main/java/com/vnb/RestaurantVote/model/Meal.java
@@ -49,7 +49,7 @@ public class Meal {
         this.name = name;
     }
 
-    public double getPrice() {
+    public Integer getPrice() {
         return price;
     }
